+++ src/include/platform/PlatformManager.h
@@ -27,6 +27,10 @@
 #include <platform/CHIPDeviceEvent.h>
 #include <system/SystemLayer.h>
 
+#if CHIP_SYSTEM_CONFIG_USE_LWIP
+#include <system/LwIPEventSupport.h>
+#endif // CHIP_SYSTEM_CONFIG_USE_LWIP
+
 namespace chip {
 
 namespace DeviceLayer {
@@ -176,7 +180,9 @@ private:
     friend class Internal::GenericThreadStackManagerImpl_OpenThread_LwIP;
     template <class>
     friend class Internal::GenericConfigurationManagerImpl;
+#if CHIP_SYSTEM_CONFIG_USE_LWIP
     friend class System::PlatformEventing;
+#endif // CHIP_SYSTEM_CONFIG_USE_LWIP
 
     /*
      * PostEvent can be called safely on any thread without locking the stack.
