+++ src/inet/IPEndPointBasis.cpp
@@ -661,7 +661,7 @@ done:
 CHIP_ERROR IPEndPointBasis::PostPacketBufferEvent(chip::System::Layer & aLayer, System::Object & aTarget,
                                                   System::EventType aEventType, System::PacketBufferHandle && aBuffer)
 {
-    const CHIP_ERROR error = aLayer.WatchableEventsManager().PostEvent(
+    const CHIP_ERROR error = aLayer.PostEvent(
         aTarget, aEventType, (uintptr_t) System::LwIPPacketBufferView::UnsafeGetLwIPpbuf(aBuffer));
     if (error == CHIP_NO_ERROR)
     {
