+++ src/platform/BUILD.gn
@@ -262,7 +262,7 @@ if (chip_device_platform != "none") {
       "Globals.cpp",
       "LockTracker.cpp",
       "PersistedStorage.cpp",
-      "SystemEventSupport.cpp",
+      "LwIPEventSupport.cpp",
       "TestIdentity.cpp",
     ]
