+++ src/platform/LwIPEventSupport.cpp
@@ -27,6 +27,8 @@
 
 #include <platform/PlatformManager.h>
 
+#if CHIP_SYSTEM_CONFIG_USE_LWIP
+
 namespace chip {
 namespace System {
 
@@ -67,3 +69,5 @@ CHIP_ERROR PlatformEventing::StartTimer(System::Layer & aLayer, uint32_t aMillis
 
 } // namespace System
 } // namespace chip
+
+#endif // CHIP_SYSTEM_CONFIG_USE_LWIP
